Separate gobuster program/command line from core functionality (libgobuster)
